requirements.txt
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
alembic==1.13.1
plotly==5.22.0
streamlit==1.35.0
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.0
jupyter==1.0.0
ipykernel==6.29.4
